refactor(optimizers): split ReferenceData builder out of objective.py
Move ReferenceValue and ReferenceData (verbatim, including their
method-local lazy imports) from q2mm/optimizers/objective.py into a new
q2mm/optimizers/reference.py so callers that only build reference data
do not pull in the heavier ObjectiveFunction machinery.

objective.py keeps ObjectiveFunction and its __getattr__ shim, and adds a
backward-compatible re-export of ReferenceData/ReferenceValue. Real
sources (io/reference.py, optimizers/__init__.py, the five evaluator
modules, and test/test_hessian_grad.py) now import from the new module.
The architecture doc module tree lists optimizers/reference.py.
